Add ConnWrapper
